Split parse-osmium into input-handler and progress-display
Rename classes and put them in separate files. parse_osmium_t gets the
new name input_handler_t and parse_stats_t gets the new name
progress_display_t.
